Client-config integration test clean-up
